+++ src/backend/parser/parse_func.c
@@ -1220,7 +1220,7 @@ setup_tlist(char *attname, Oid relid)
 	Resdom	   *resnode;
 	Var		   *varnode;
 	Oid			typeid;
-	int16		type_mod;
+	int32		type_mod;
 	int			attno;
 
 	attno = get_attnum(relid, attname);
