+++ src/include/nodes/parsenodes.h
@@ -644,7 +644,7 @@ typedef struct TypeName
 	char	   *name;			/* name of the type */
 	bool		timezone;		/* timezone specified? */
 	bool		setof;			/* is a set? */
-	int16		typmod;			/* type modifier */
+	int32		typmod;			/* type modifier */
 	List	   *arrayBounds;	/* array bounds */
 } TypeName;
